Add files via upload
This code presents a water tank system that simulates water usage using abstract classes.  Home Tank (200L) and Building Tank (1000L) are the two subclasses from which the user can select. The user can also choose to fill, use the water, and check the status.  Additionally, JOptionPane is used for input and messages in this code, which terminates immediately if the tank is full or empty.  Additionally, by defining common methods in the abstract class and implementing them in particular tank classes, the code makes use of OOP (Object-Oriented Program) features, such as abstraction, inheritance, and even polymorphism.